PB/Backend-TFC/requirements.txt
# Requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
Pillow==9.3.0
platformdirs==2.5.4
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
virtualenv==20.17.0
